Move item serializers to own file
