+++ app/stac_api/views/item.py
@@ -14,8 +14,8 @@ from rest_framework_condition import etag
 from stac_api.models import Asset
 from stac_api.models import Collection
 from stac_api.models import Item
-from stac_api.serializers.serializers import AssetSerializer
-from stac_api.serializers.serializers import ItemSerializer
+from stac_api.serializers.item import AssetSerializer
+from stac_api.serializers.item import ItemSerializer
 from stac_api.serializers.serializers_utils import get_relation_links
 from stac_api.utils import get_asset_path
 from stac_api.utils import utc_aware
