+++ app/tests/tests_09/test_serializer.py
@@ -14,8 +14,8 @@ from rest_framework.test import APIRequestFactory
 
 from stac_api.models import get_asset_path
 from stac_api.serializers.collection import CollectionSerializer
-from stac_api.serializers.serializers import AssetSerializer
-from stac_api.serializers.serializers import ItemSerializer
+from stac_api.serializers.item import AssetSerializer
+from stac_api.serializers.item import ItemSerializer
 from stac_api.utils import get_link
 from stac_api.utils import isoformat
 from stac_api.utils import utc_aware
